git-tfs-id: [http://developmentvlan:8585/tfs/vanrise.collection]$/;C26165

+++ CDRAnalysis/Vanrise.Fzero.FraudAnalysis.Entities/BTSHighValueCases.cs
@@ -6,7 +6,7 @@ namespace Vanrise.Fzero.FraudAnalysis.Entities
     {
         public int? BTS_Id { get; set; }
 
-        public int Volume { get; set; }
+        public decimal Volume { get; set; }
 
     }
